compare offtargets

+++ beditor/lib/get_specificity.py
@@ -103,8 +103,9 @@ def dguides2offtargets(cfg):
 #     genomegffp='{}/{}/{}.{}.{}.gff3.gz'.format(genomeannotd,cfg['host'],host_,cfg['genomeassembly'],cfg['genomerelease'])
 
     datatmpd='{}/tmp'.format(cfg['datad'],stepn)
-    for dp in [datatmpd]: 
-        makedirs(dp,exist_ok=cfg['force'])
+    for dp in [datatmpd]:
+        if not exists(dp):
+            makedirs(dp)
 
     dguides=pd.read_csv(dguidesp)
 #     dguides.to_csv('{}/{}'.format(cfg['datad'],basename(dguidesp)),sep='\t')
@@ -115,9 +116,10 @@ def dguides2offtargets(cfg):
     otBedFname = batchBase+".bed"
     print(otBedFname)
     faFname = batchBase+".fa"
-    with open(faFname,'w') as f:
-        for gi in dguides.index:
-            f.write('>{}\n{}\n'.format(gi,dguides.loc[gi,'guide sequence+PAM']))
+    if not exists(faFname) or cfg['force']:
+        with open(faFname,'w') as f:
+            for gi in dguides.index:
+                f.write('>{}\n{}\n'.format(gi.replace(' ','_'),dguides.loc[gi,'guide sequence+PAM']))
 
     # BWA: allow up to X mismatches
     maxMMs=5
@@ -152,7 +154,8 @@ def dguides2offtargets(cfg):
     samp = batchBase+".sam"
     genomep=cfg['genomep']
     genomed = dirname(genomep) # make var local, see below
-
+    genomegffp=cfg['genomegffp']
+    
     open(matchesBedFname, "w") # truncate to 0 size
 
     # increase MAXOCC if there is only a single query, but only in CGI mode
@@ -160,98 +163,130 @@ def dguides2offtargets(cfg):
     seqLen = guidel
 
     bwaM = MFAC*MAXOCC # -m is queue size in bwa
-    cmd = "bwa aln -o 0 -m %(bwaM)s -n %(maxDiff)d -k %(maxDiff)d -N -l %(seqLen)d %(genomep)s %(faFname)s > %(saFname)s" % locals()
-    runbashcmd(cmd)
-
-    cmd = "bwa samse -n %(MAXOCC)d %(genomep)s %(saFname)s %(faFname)s > %(samp)s" % locals()
-    runbashcmd(cmd)
+    if not exists(saFname) or cfg['force']:
+        cmd = "bwa aln -o 0 -m %(bwaM)s -n %(maxDiff)d -k %(maxDiff)d -N -l %(seqLen)d %(genomep)s %(faFname)s > %(saFname)s" % locals()
+        runbashcmd(cmd)
+        
+    if not exists(samp) or cfg['force']:
+        cmd = "bwa samse -n %(MAXOCC)d %(genomep)s %(saFname)s %(faFname)s > %(samp)s" % locals()
+        runbashcmd(cmd)
+    
     #----make tables-----------
     gff_colns = ['chromosome', 'source', 'type', 'start', 'end', 'score', 'strand', 'phase', 'attributes']
     bed_colns = ['chromosome','start','end','id','NM','strand']
-
-    samfile=pysam.AlignmentFile(samp, "rb")
-    dalignbed=pd.DataFrame(columns=bed_colns)
-    for read in samfile.fetch():
-        algnids=[]
-        algnids.append('{}|{}{}|{}|{}'.format(read.reference_name,
-                 ('-' if read.is_reverse else '+'),read.positions[0],read.cigarstring,read.get_tag('NM')))
-        algnids+=['|'.join(s.split(',')) for s in read.get_tag('XA').split(';') if len(s.split(','))>1]
-        chroms=[]
-        starts=[]
-        ends=[]
-        ids=algnids
-        NMs=[]
-        strands=[]    
-        for a in algnids:
-            chroms.append(a.split('|')[0])
-            starts.append(int(a.split('|')[1][1:])-1)
-            ends.append(int(a.split('|')[1][1:])+str2num(a.split('|')[2])-1)
-            NMs.append(a.split('|')[3])
-            strands.append(a.split('|')[1][0])
-        col2dalignbed={'chromosome':chroms,
-                       'start':starts,
-                       'end':ends,
-                       'id':ids,
-                       'NM':NMs,
-                       'strand':strands}
-    #     col2dalignbed=dict(zip(cols,[a.split('|')[0],a.split('|')[1],a.split('|')[2],a,a.split('|')[3],a.split('|')[4] for a in algnids]))
-        dalignbed_=pd.DataFrame(col2dalignbed)
-        dalignbed_['guide: id']=read.qname
-        dalignbed = dalignbed.append(dalignbed_,ignore_index=True)
-    #     break
-    samfile.close()
     
-    dalignbedp='{}/dalignbed.tsv'.format(datatmpd)
-    dalignbed.to_csv(dalignbedp,sep='\t')
-
     alignmentbedp='{}/alignment.bed'.format(datatmpd)
-    dalignbed.loc[:,bed_colns].to_csv(alignmentbedp,sep='\t',
-                    header=False,index=False)
+    dalignbedp='{}/dalignbed.tsv'.format(datatmpd)
+    if not exists(alignmentbedp) or cfg['force']:
+        samfile=pysam.AlignmentFile(samp, "rb")
+        dalignbed=pd.DataFrame(columns=bed_colns)
+        for read in samfile.fetch():
+            algnids=[]
+            algnids.append('{}|{}{}|{}|{}'.format(read.reference_name,
+                     ('-' if read.is_reverse else '+'),read.positions[0],read.cigarstring,read.get_tag('NM')))
+            if read.has_tag('XA'):
+                algnids+=['|'.join(s.split(',')) for s in read.get_tag('XA').split(';') if len(s.split(','))>1]
+            chroms=[]
+            starts=[]
+            ends=[]
+            ids=algnids
+            NMs=[]
+            strands=[]    
+            for a in algnids:
+                chroms.append(a.split('|')[0])
+                starts.append(int(a.split('|')[1][1:])-1)
+                ends.append(int(a.split('|')[1][1:])+str2num(a.split('|')[2])-1)
+                NMs.append(a.split('|')[3])
+                strands.append(a.split('|')[1][0])
+            col2dalignbed={'chromosome':chroms,
+                           'start':starts,
+                           'end':ends,
+                           'id':ids,
+                           'NM':NMs,
+                           'strand':strands}
+        #     col2dalignbed=dict(zip(cols,[a.split('|')[0],a.split('|')[1],a.split('|')[2],a,a.split('|')[3],a.split('|')[4] for a in algnids]))
+            dalignbed_=pd.DataFrame(col2dalignbed)
+            dalignbed_['guide: id']=read.qname.replace('_',' ')
+            dalignbed = dalignbed.append(dalignbed_,ignore_index=True)
+        #     break
+        samfile.close()
+
+        # filter bad asssembly junk genmomes
+        from beditor.lib.global_vars import host2contigs
+        dalignbed=dalignbed.loc[dalignbed['chromosome'].isin(host2contigs[cfg['host']]),:]
+        dalignbed.to_csv(dalignbedp,sep='\t')
+        dalignbed.loc[:,bed_colns].to_csv(alignmentbedp,sep='\t',
+                        header=False,index=False)
+    else:
+        dalignbed=pd.read_csv(dalignbedp,sep='\t')
+        dalignbed=dalignbed.drop([c for c in dalignbed if 'Unnamed' in c],axis=1)
+        
     alignmentbedsortedp=alignmentbedp+'.sorted.bed'
-    cmd='bedtools sort -i {} > {}'.format(alignmentbedp,alignmentbedsortedp)
-    runbashcmd(cmd)
-    genomegffsortedp=genomegffp+'.sorted.gff3.gz'
-    cmd='bedtools sort -i {} > {}'.format(genomegffp,genomegffsortedp)
-    runbashcmd(cmd)
-    annotationsbedp='{}/annotations.bed'.format(datatmpd)
-    cmd='bedtools intersect -wa -wb -loj -a {} -b {} > {}'.format(alignmentbedsortedp,genomegffsortedp,annotationsbedp)
-    runbashcmd(cmd)
-#     if the error in human, use: `cut -f 1 data/alignment.bed.sorted.bed | sort| uniq -c | grep -v CHR | grep -v GL | grep -v KI`
-
-    dannots=pd.read_csv('{}/annotations.bed'.format(datatmpd),sep='\t',
-               names=bed_colns+gff_colns)
-
-    dalignbed=dalignbed.set_index('grnaId').join(dguides)
-    dalignbed=dalignbed.reset_index().set_index('id')
-
-    dalignbed.to_csv('data/dalignbedguides.tsv','\t')
+    if not exists(alignmentbedsortedp) or cfg['force']:
+        cmd='bedtools sort -i {} > {}'.format(alignmentbedp,alignmentbedsortedp)
+        runbashcmd(cmd)
     
+    genomegffsortedp=genomegffp+'.sorted.gff3.gz'
+    if not exists(genomegffsortedp) or cfg['force']:    
+        cmd='bedtools sort -i {} > {}'.format(genomegffp,genomegffsortedp)
+        runbashcmd(cmd)
+        
+    annotationsbedp='{}/annotations.bed'.format(datatmpd)
+    if not exists(annotationsbedp) or cfg['force']:    
+        cmd='bedtools intersect -wa -wb -loj -a {} -b {} > {}'.format(alignmentbedsortedp,genomegffsortedp,annotationsbedp)
+        runbashcmd(cmd)
+#     if the error in human, use: `cut -f 1 data/alignment.bed.sorted.bed | sort| uniq -c | grep -v CHR | grep -v GL | grep -v KI`
+    dalignbedguidesp='{}/dalignbedguides.tsv'.format(datatmpd)
+    if not exists(dalignbedguidesp) or cfg['force']:
+        dalignbed=dalignbed.set_index('guide: id').join(dguides)
+        dalignbed=dalignbed.reset_index().set_index('guide: id')
+        dalignbed.to_csv(dalignbedguidesp,'\t')
+    else:
+        dalignbed=pd.read_csv(dalignbedguidesp,'\t')
+        dalignbed=dalignbed.drop([c for c in dalignbed if 'Unnamed' in c],axis=1)
+        
     # dalignid2seq=pd.DataFrame(columns=['sequence'])
     # dalignid2seq.index.name='id'
-    alignedfastap='{}/alignment.fa'.format(datatmpd)
-    cmd='bedtools getfasta -s -name -fi {} -bed {} -fo {}'.format(genomep,alignmentbedp,alignedfastap)
-    runbashcmd(cmd)
-    
-    dalignedfasta=fa2df(alignedfastap)
-    dalignedfasta.columns=['aligned sequence']
+    dalignedfastap='{}/dalignedfasta.tsv'.format(datatmpd)
+    if not exists(dalignedfastap) or cfg['force']:
+        alignedfastap='{}/alignment.fa'.format(datatmpd)
+        if not exists(alignedfastap) or cfg['force']:
+            cmd='bedtools getfasta -s -name -fi {} -bed {} -fo {}'.format(genomep,alignmentbedp,alignedfastap)
+            runbashcmd(cmd)
 
-    dalignbed=dalignbed.set_index('id').join(dalignedfasta)
-    
+        dalignedfasta=fa2df(alignedfastap)
+        dalignedfasta.columns=['aligned sequence']
+        dalignedfasta.to_csv(dalignedfastap,sep='\t')
+    else:
+        dalignedfasta=pd.read_csv(dalignedfastap,sep='\t')        
+        dalignedfasta=dalignedfasta.drop([c for c in dalignbed if 'Unnamed' in c],axis=1)
+        
     dalignbedguidesseqp='{}/dalignbedguidesseq.tsv'.format(datatmpd)
-    dalignbed.to_csv(dalignbedguidesseqp,sep='\t')
-    # for seq in SeqIO.parse(alignedfastap,"fasta"):
-    #     dalignbed.loc[seq.id,'aligned sequence']=str(seq.seq.upper())
-    # #     break
-    dalignbed=dalignbed.drop_duplicates()
-
+    if not exists(dalignbedguidesseqp) or cfg['force']:        
+#         from beditor.lib.io_dfs import df2info
+#         if cfg['test']:
+#             df2info(dalignbed)
+#             df2info(dalignedfasta)
+            
+#         dalignedfasta['guide: id']=dalignedfasta.apply(lambda x : x['guide: id'].replace('_',' '),axis=1)
+        dalignbed=dalignbed.reset_index().set_index('id').join(dalignedfasta)
+        dalignbed.index.name='id'
+        dalignbed=dalignbed.drop_duplicates()
+        dalignbed.to_csv(dalignbedguidesseqp,sep='\t')
+    else:
+        dalignbed=pd.read_csv(dalignbedguidesseqp,sep='\t')
+        dalignbed=dalignbed.drop([c for c in dalignbed if 'Unnamed' in c],axis=1)
+                
     dalignbed['Hamming distance']=dalignbed.apply(lambda x : hamming_distance(x['guide sequence+PAM'], x['aligned sequence']),axis=1)
     dalignbed['alignment','alignment: score']=dalignbed.apply(lambda x: align(x['guide sequence+PAM'],x['aligned sequence'],l=guidel),axis=1)
 
     # for i in dalignbed.index:
     #     dalignbed.loc[i,'alignment'],dalignbed.loc[i,'alignment: score']=align(dalignbed.loc[i,'guide sequence+PAM'],dalignbed.loc[i,'aligned sequence'],l=guidel)
+    dannots=pd.read_csv('{}/annotations.bed'.format(datatmpd),sep='\t',
+               names=bed_colns+gff_colns)
 
     dcombo=dalignbed.join(dannots.set_index('id'),rsuffix='.2')
-    dcombo.to_csv('{}/dcombo.tsv'.format(cfg['datad']),sep='\t')
+    dcombo.to_csv('{}/dofftargets.tsv'.format(cfg['datad']),sep='\t')
 
 
 # if __name__ == '__main__':
